Move some tests outside EDT
Since 2021.2, commit 02e3b78, tests execution needs access to the Event Dispatch Thread. Unfortunately,
we have some tests that do test IntelliJ test execution. All the test, by default, are run in EDT. In
order to fix it, we need to move this tests outside EDT and fix errors related to resources that may
be accessed only from EDT

https://github.com/JetBrains/intellij-community/commit/02e3b78d485ea9ba8459d4f96b26e56c79a49eba#diff-3279fab1ac7d401ffbecaba99be6af3ddab973a8d1750b806ff7d1e35bf53c9aL244-L245
